refactor: split helper file functions
refactor ts/helpers/keyboard-events.ts and ts/helpers/media-breakpoints.ts to isolate those functions using React, in an effort to compensate for Webpack's inability to tree-shake these out of the dotcomchrome bundle.
