+++ apps/site/assets/ts/components/ExpandableBlock.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement, useState } from "react";
-import { handleReactEnterKeyPress } from "../helpers/keyboard-events";
+import { handleReactEnterKeyPress } from "../helpers/keyboard-events-react";
 import renderSvg from "../helpers/render-svg";
 import renderFa from "../helpers/render-fa";
 import { caret } from "../helpers/icon";
