+++ apps/site/assets/ts/helpers/keyboard-events.ts
@@ -1,5 +1,3 @@
-import { KeyboardEvent as ReactKeyboardEvent } from "react";
-
 export const isEnter = (key: number | string): boolean =>
   key === "Enter" || key === 13;
 
@@ -16,22 +14,3 @@ export const handleNativeEscapeKeyPress = (
   e: KeyboardEvent,
   cb: Function
 ): void => (isEscape(e.key || e.keyCode) ? cb(e) : () => {});
-
-// for React
-export const handleReactEnterKeyPress = (
-  e: ReactKeyboardEvent,
-  onClick: Function
-): void => {
-  if (e.key === "Enter") {
-    onClick();
-  }
-};
-
-export const handleReactExitKeyPress = (
-  e: ReactKeyboardEvent,
-  onClick: Function
-): void => {
-  if (e.key === "Escape") {
-    onClick();
-  }
-};
