+++ apps/site/assets/ts/helpers/media-breakpoints.ts
@@ -1,4 +1,3 @@
-import { useEffect, useState } from "react";
 // Specify in pixels, but in numbers here for calculations
 const mediaBreakpoints = {
   xs: 0,
@@ -15,16 +14,3 @@ export const isSMDown = (): boolean =>
   typeof document !== "undefined"
     ? window.matchMedia(`(max-width: ${mediaBreakpoints.sm}px)`).matches
     : false;
-
-export const useSMDown = (): boolean => {
-  const [isSmallBreakpoint, setIsSmallBreakpoint] = useState(isSMDown());
-  useEffect(() => {
-    const handleResize = (): void => {
-      setIsSmallBreakpoint(isSMDown());
-    };
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
-
-  return isSmallBreakpoint;
-};
