+++ apps/site/assets/ts/schedule/components/schedule-finder/SelectContainer.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement, ReactNode } from "react";
-import { handleReactEnterKeyPress } from "../../../helpers/keyboard-events";
+import { handleReactEnterKeyPress } from "../../../helpers/keyboard-events-react";
 import renderSvg from "../../../helpers/render-svg";
 import arrowIcon from "../../../../static/images/icon-down-arrow.svg";
